pyproject.toml
[tool.poetry]
name = "interviews"
version = "0.1.0"
description = ""
authors = ["Mikhail <mixail.critsyn@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.79.0"
tortoise-orm = "^0.19.2"
aerich = "^0.6.3"
psycopg2-binary = "^2.9.3"
asyncpg = "^0.26.0"
PyJWT = "^2.4.0"
passlib = "^1.7.4"
aiofiles = "^0.8.0"
uvicorn = "^0.18.2"
pydantic = {extras = ["email"], version = "^1.9.1"}
emails = "^0.6"
python-multipart = "^0.0.5"
itsdangerous = "^2.1.2"
bcrypt = "^3.2.2"
Jinja2 = "^3.1.2"
fastapi-pagination = "^0.9.3"
google-auth = "^2.9.1"
redis = "^4.3.4"
celery = "^5.2.7"
Authlib = "^1.0.1"

[tool.poetry.dev-dependencies]


[tool.aerich]
tortoise_orm = "src.config.tortoise_config.TORTOISE_ORM"
location = "src/db/migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
